@@ -1,101 +1,62 @@
-import { useEffect } from 'react';
+import { KeyboardControls } from "@react-three/drei";
+import { Canvas } from "@react-three/fiber";
+import { Physics } from "@react-three/rapier";
+import { Suspense, useEffect, useMemo } from "react";
+import { Experience } from "./components/Experience";
+import { UI } from "./components/UI";
+import { Stats } from '@react-three/drei'
+import { Perf } from "r3f-perf";
 
-import * as THREE from 'three';
-// import { FBXLoader } from 'three/examples/jsm/loaders/FBXLoader';
-// import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader';
-// import { VOXLoader } from 'three/examples/jsm/loaders/VOXLoader';
-import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
 
-import SceneInit from './lib/SceneInit';
-import { init } from './scripts/CameraControls';
-import CharacterControllerDemo from './lib/CharacterControls';
-import UserInput from './UserInputs.js';
-
+export const Controls = {
+  forward: "forward",
+  back: "back",
+  left: "left",
+  right: "right",
+  jump: "jump",
+};
 
 function App() {
-  let input = new UserInput();
-
-  useEffect(() => {
-
-
-    const test = new SceneInit('myThreeJsCanvas', input);
-    test.initialize();
-    test.animate();
-
-    // const boxGeometry = new THREE.BoxGeometry(8, 8, 8);
-    // const boxMaterial = new THREE.MeshNormalMaterial();
-    // const boxMesh = new THREE.Mesh(boxGeometry, boxMaterial);
-    // test.scene.add(boxMesh);
+  const map = useMemo(
+    () => [
+      { name: Controls.forward, keys: ["ArrowUp", "KeyW"] },
+      { name: Controls.back, keys: ["ArrowDown", "KeyS"] },
+      { name: Controls.left, keys: ["ArrowLeft", "KeyA"] },
+      { name: Controls.right, keys: ["ArrowRight", "KeyD"] },
+      { name: Controls.jump, keys: ["Space"] },
+    ],
+    []
+  );
 
 
-
-
-    let loadedModel;
-    const glftLoader = new GLTFLoader();
-    glftLoader.load('./assets/shiba/scene.gltf', (gltfScene) => {
-      loadedModel = gltfScene;
-      // console.log(loadedModel);
-
-      gltfScene.scene.rotation.y = Math.PI / 8;
-      gltfScene.scene.position.y = 5;
-      gltfScene.scene.scale.set(10, 10, 10);
-      test.scene.add(gltfScene.scene);
-
-
-    });
+  // useEffect(() => {
 
 
 
 
-    const soldierModel = new GLTFLoader();
-    soldierModel.load('./assets/soldier/Soldier.glb', (gltfScene) => {
-      const model = gltfScene.scene;
-      model.traverse(function (object) {
-        if (object.isMesh) object.castShadow = true;
-      });
-      test.scene.add(model);
 
 
-      const gltfAnimations = gltfScene.animations;
-      const mixer = new THREE.AnimationMixer(loadedModel);
-      const animationsMap = new Map()
-      gltfAnimations.filter(a => a.name != 'TPose').forEach((a) => {
-        animationsMap.set(a.name, mixer.clipAction(a))
-      })
+  // }, [])
 
-      //characterControls = new CharacterControls(soldierModel, mixer, animationsMap, orbitControls, camera, 'Idle')
+  return (<>
 
-    })
+    <KeyboardControls map={map}>
+      <Canvas shadows camera={{ position: [10, 10, 10], fov: 60 }}>
+        <color attach="background" args={["#ececec"]} />
+        <Suspense>
+          <Physics debug>
+            <Experience />
+          </Physics>
+        </Suspense>
+        <axesHelper args={[15]} position={[0, 0, 0]} />
+        <gridHelper args={[20]} />
+        <Stats />
+        <Perf position="bottom-left" />
+      </Canvas>
+    </KeyboardControls>
 
-    const gridHelper = new THREE.GridHelper(12, 12);
-    test.scene.add(gridHelper);
+  </>
 
-    // Sets the x, y, and z axes with each having a length of 4
-    const axesHelper = new THREE.AxesHelper(12);
-    test.scene.add(axesHelper);
-
-
-    //let CharacterControls = new CharacterControllerDemo(test.camera, test.scene);
-
-
-    const animate = () => {
-      if (loadedModel) {
-        loadedModel.scene.rotation.x += 0.01;
-        loadedModel.scene.rotation.y += 0.01;
-        loadedModel.scene.rotation.z += 0.01;
-      }
-      requestAnimationFrame(animate);
-    };
-    animate();
-  }, []);
-
-
-
-
-  return (
-    <div>
-      <canvas id="myThreeJsCanvas" />
-    </div>
   );
 }
 
